GHC 9.12 compat, update bounds
Refactor annotation utils. Exceptions have been retooled, as of
GHC 9.12. In particular, the exception context is no longer part of
the SomeException data type; the context is a separate type that is
automagically included.

This means that the displayX functions in annotation utils are useless,
since displayException only ever acts on the exception itself, not the
extra context.

We could try to use cpp to preserve them, but that is a massive
headache, so instead we remove them. Where do callstacks now show up?
Only when you specifically ask for them, which -- by default -- is
the uncaught exception handler. Thus our logic is now moved to
augmenting the handler.

To sum up, the annotation utils now offers functions to ignore known
callstacks in the handlers.